+++ CHANGELOG.md
@@ -6,16 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-+ DAVAMS-320 : Rebrand 'Klarna' to 'Klarna - Pay later in 14 days' inclusive a new logo
+### Improvements 
 + PLGOPNS-256: Use a link to documentation, instead a tooltip, in Second Chance field
 + PLGOPNS-283: [chore] Update modman file and remove comments from the lines related with test files
 + PLGOPNS-282: [test] Add second parameter to GenerateOrderData function
 + PLGOPNS-281: [test] Add OC_VERSION env variable to phpunit.xml.dist
 + PLGOPNS-284: [refactor] Fix placeholder for 'lifetime payment link' field
 
+### Fixed
++ PLGOPNS-295: Add getEventByCode function, which does not exist prior to OC 3.0.2
+
+
 ## 3.1.0
 Release date: September, 14th 2020
 
+### Improvements 
 + PLGOPNS-246: Add a notification message if PHP version is not supported by the plugin.
 + PLGOPNS-243: Improve fields related with lifetime of the payment link in the settings
 + PLGOPNS-253: Validate the field lifetime of payment link to avoid negative number